+++ .idea/encodings.xml
@@ -20,6 +20,13 @@
     <file url="file://$PROJECT_DIR$/Lesson45/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/Lesson45SpringDataJpa/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/Lesson46SpringBootE-store/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Lesson47/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Lesson47/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Lesson48/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Lesson48/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Lesson48SoapClient/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Lesson48SoapClient/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Lesson48SoapClient/target/generated-sources/wsimport" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
   </component>
