+++ Lesson45SpringDataJpa/src/main/java/us/teachmeskills/springdataJpa/app/Lesson45SpringDataJpaApplication.java
@@ -8,10 +8,10 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @SpringBootApplication
 @EnableJpaRepositories("us.teachmeskills.springdataJpa.repositories")
 public class Lesson45SpringDataJpaApplication {
-    private final Environment environment;
+    private final Environment env;
 
-    public Lesson45SpringDataJpaApplication(Environment environment) {
-        this.environment = environment;
+    public Lesson45SpringDataJpaApplication(Environment env) {
+        this.env = env;
     }
 
     public static void main(String[] args) {
