+++ Lesson45SpringDataJpa/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/c18-onl2022
+spring.datasource.url=jdbc:postgresql://localhost:5432/c18-onl2022/test
 spring.datasource.username=postgres
 spring.datasource.password=Barsukov28071992
 spring.jpa.show-sql=true
