+++ Lesson48/src/main/java/META-INF/JAXB/episode_xjc_1.xjb
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<bindings version="2.1" xmlns="http://java.sun.com/xml/ns/jaxb">
+    <!--
+
+  This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2
+  See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a>
+  Any modifications to this file will be lost upon recompilation of the source schema.
+  Generated on: 2023.05.23 at 11:38:43 PM EDT
+
+    -->
+    <bindings scd="x-schema::tns" xmlns:tns="http://www.baeldung.com/springsoap/gen">
+        <schemaBindings map="false">
+            <package name="by.teachmeskills.soap.model"/>
+        </schemaBindings>
+        <bindings scd="tns:getCountryRequest">
+            <class ref="by.teachmeskills.soap.model.GetCountryRequest"/>
+        </bindings>
+        <bindings scd="tns:getCountryResponse">
+            <class ref="by.teachmeskills.soap.model.GetCountryResponse"/>
+        </bindings>
+        <bindings scd="~tns:country">
+            <class ref="by.teachmeskills.soap.model.Country"/>
+        </bindings>
+        <bindings scd="~tns:currency">
+            <typesafeEnumClass ref="by.teachmeskills.soap.model.Currency"/>
+        </bindings>
+    </bindings>
+</bindings>
+
