--- Lesson48/src/main/java/by/teachmeskills/Application.java
@@ -1,11 +1,0 @@
-package by.teachmeskills;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
