+++ Lesson48/src/main/java/by/teachmeskills/soap/model/ObjectFactory.java
@@ -2,7 +2,7 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2022.07.02 at 05:13:43 PM MSK 
+// Generated on: 2023.05.23 at 11:38:43 PM EDT 
 //
