+++ Lesson48SoapClient/src/main/java/by/teachmeskills/Application.java
