+++ Lesson48SoapClient/src/main/java/by/teachmeskills/Main.java
