+++ Lesson48SoapClient/src/main/resources/application.properties
