+++ Lesson48SoapClient/src/main/resources/countries.wsdl
